Fix ClownCar_IV: revert state dtype cast — only change is SKIP_GPTQ=1
new_state.to(dtype) cast state from float32 → bfloat16 during training,
destroying delta rule accumulator precision. Caused 0.27 BPB regression.
ClownCar_IV is now identical to ClownCar_II except SKIP_GPTQ=1.

+++ experiments/ClownCar_IV/train_gpt.py
@@ -1397,7 +1397,7 @@ class CanonicalDeltaNet(nn.Module):
             output_final_state=True,
         )
         y = o.reshape(B, T, H * Dh)
-        return self.norm(x + self.o_proj(y)), new_state.to(dtype)
+        return self.norm(x + self.o_proj(y)), new_state
 
 
 # flat blocks (unique, U-Net enc/dec) + crawler blocks (shared, looped K times)
